pyproject.toml
[project]
name = "vidgen"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "edge-tts>=6.1.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "pygments>=2.17.0",
    "playwright>=1.40.0",
    "loguru>=0.7.0",
]
